<add> Adding modals with react-modal to be used on update forms, add forms, delete forms. <change> changing the functionality of logout process, to be called when the backend cookie is expired but the frontend token is active, this makes the user logout and perform login again

+++ src/actions/cartRequests.js
@@ -1,19 +1,27 @@
 'use server'
 
 import { revalidatePath } from 'next/cache';
-import { cookies } from "next/headers";
+import { cookieFetchVerification } from '@/lib/cookieVerification';
 
 export async function fetchCartInfoByUser(){
-    console.log('FETCHING CART INFO BY USER')
-    const allCookies = cookies();
-    const connectSidCookie = allCookies.getAll('connect.sid');
-    const cookieForServer = `${connectSidCookie[0].name}=${connectSidCookie[0].value}`
+    console.log('CALLING FETCH CART INFO')
+    const { cookieForServer, expired } = await cookieFetchVerification();
+
+    if (expired) {
+        console.log('Session expired on the backend. Triggering logout.');
+        return { expired: true };
+    }
+
     try {
         const response = await fetch(`${process.env.NEXT_PUBLIC_BACKEND_URL}/cart`, {
             headers : {cookie: cookieForServer}
         })
 
-        if (!response.ok) {       
+        if (!response.ok) { 
+            if (response.status === 401 || response.status === 403) {
+                console.log('Session expired on the backend. Triggering logout.');
+                return { expired: true };
+            }      
             const errorResponse = await response.json();
             console.log(`FETCHING CART INFO FAILED`, errorResponse);
             throw new Error(`Error ${errorResponse.status}: ${errorResponse?.customError?.message || errorResponse.error}`);
@@ -32,9 +40,12 @@ export async function fetchCartInfoByUser(){
 
 export async function updateCartItem({product_id, qty}){
     console.log('UPDATE CART ITEM FETCH', product_id, qty)
-    const allCookies = cookies();
-    const connectSidCookie = allCookies.getAll('connect.sid');
-    const cookieForServer = `${connectSidCookie[0].name}=${connectSidCookie[0].value}`
+    const { cookieForServer, expired } = await cookieFetchVerification();
+
+    if (expired) {
+        console.log('Session expired on the backend. Triggering logout.');
+        return { expired: true };
+    }
     try {
         const response = await fetch(`${process.env.NEXT_PUBLIC_BACKEND_URL}/cart`, {
             method: 'PATCH',
@@ -48,7 +59,11 @@ export async function updateCartItem({product_id, qty}){
             }
         })
 
-        if (!response.ok) {       
+        if (!response.ok) {
+            if (response.status === 401 || response.status === 403) {
+                console.log('Session expired on the backend. Triggering logout.');
+                return { expired: true };
+            }       
             const errorResponse = await response.json();
             console.log(`UPDATING CART ITEM FAILED`, errorResponse);
             throw new Error(`Error ${errorResponse.status}: ${errorResponse?.customError?.message || errorResponse.error}`);
@@ -66,9 +81,13 @@ export async function updateCartItem({product_id, qty}){
 
 export async function deleteCartItem(id){
     console.log('DELETE CART ITEM ID FETCH:', id)
-    const allCookies = cookies();
-    const connectSidCookie = allCookies.getAll('connect.sid');
-    const cookieForServer = `${connectSidCookie[0].name}=${connectSidCookie[0].value}`
+    const { cookieForServer, expired } = await cookieFetchVerification();
+
+    if (expired) {
+        console.log('Session expired on the backend. Triggering logout.');
+        return { expired: true };
+    }
+    
     try {
         const response = await fetch(`${process.env.NEXT_PUBLIC_BACKEND_URL}/cart/${id}`, {
             method: 'DELETE',
@@ -77,7 +96,11 @@ export async function deleteCartItem(id){
             }
         })
 
-        if (!response.ok) {       
+        if (!response.ok) {
+            if (response.status === 401 || response.status === 403) {
+                console.log('Session expired on the backend. Triggering logout.');
+                return { expired: true };
+            }       
             const errorResponse = await response.json();
             console.log(`DELETE CART ITEM FAILED`, errorResponse);
             throw new Error(`Error ${errorResponse.status}: ${errorResponse?.customError?.message || errorResponse.error}`);
